requirements.txt
openai>=1.0.0
PyMuPDF==1.25.3
moviepy==1.0.3
edge-tts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
